refactor: 코드 패키징 변경
